refactor: apply review suggestions
