+++ frontned/src/contants/constants.ts
@@ -1,0 +1,1 @@
+export const STALE_TIME_MIN = 5 * 60 * 1000;
