+++ frontned/src/providers/routes.tsx
@@ -1,0 +1,11 @@
+import { createBrowserRouter } from 'react-router-dom';
+import HomePage from '../pages/HomePage/HomePage';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <HomePage />,
+  },
+]);
+
+export default router;
